+++ examples/OpenGL3/OpenGL3_TexturedSphere.c
@@ -66,16 +66,16 @@ void GenerateSphere(float radius, int sectors, int stacks)
     /* Generate vertices */
     for (i = 0; i <= stacks; i++)
     {
-        float phi = (float)i / (float)stacks * M_PI;
-        float sinPhi = sin(phi);
-        float cosPhi = cos(phi);
+        float phi = (float)i / (float)stacks * (float)M_PI;
+        float sinPhi = (float)sin(phi);
+        float cosPhi = (float)cos(phi);
         float v = 1.0f - (float)i / (float)stacks;
 
         for (j = 0; j <= sectors; j++)
         {
-            float theta = (float)j / (float)sectors * 2.0f * M_PI;
-            float sinTheta = sin(theta);
-            float cosTheta = cos(theta);
+            float theta = (float)j / (float)sectors * 2.0f * (float)M_PI;
+            float sinTheta = (float)sin(theta);
+            float cosTheta = (float)cos(theta);
             float u = (float)j / (float)sectors;
 
             float x = radius * sinPhi * cosTheta;
@@ -229,8 +229,8 @@ void MatrixIdentity(float* m)
 
 void MatrixRotateY(float* m, float angleDeg)
 {
-    float rad = angleDeg * M_PI / 180.0f;
-    float c = cos(rad), s = sin(rad);
+    float rad = angleDeg * (float)M_PI / 180.0f;
+    float c = (float)cos(rad), s = (float)sin(rad);
     MatrixIdentity(m);
     m[0] = c; m[2] = s;
     m[8] = -s; m[10] = c;
@@ -238,8 +238,8 @@ void MatrixRotateY(float* m, float angleDeg)
 
 void MatrixRotateX(float* m, float angleDeg)
 {
-    float rad = angleDeg * M_PI / 180.0f;
-    float c = cos(rad), s = sin(rad);
+    float rad = angleDeg * (float)M_PI / 180.0f;
+    float c = (float)cos(rad), s = (float)sin(rad);
     MatrixIdentity(m);
     m[5] = c; m[6] = s;
     m[9] = -s; m[10] = c;
@@ -267,13 +267,13 @@ void MatrixLookAt(float* m, float eyeX, float eyeY, float eyeZ,
     f[0] = centerX - eyeX;
     f[1] = centerY - eyeY;
     f[2] = centerZ - eyeZ;
-    len = sqrt(f[0] * f[0] + f[1] * f[1] + f[2] * f[2]);
+    len = (float)sqrt(f[0] * f[0] + f[1] * f[1] + f[2] * f[2]);
     if (len != 0.0f) { f[0] /= len; f[1] /= len; f[2] /= len; }
 
     s[0] = f[1] * upZ - f[2] * upY;
     s[1] = f[2] * upX - f[0] * upZ;
     s[2] = f[0] * upY - f[1] * upX;
-    len = sqrt(s[0] * s[0] + s[1] * s[1] + s[2] * s[2]);
+    len = (float)sqrt(s[0] * s[0] + s[1] * s[1] + s[2] * s[2]);
     if (len != 0.0f) { s[0] /= len; s[1] /= len; s[2] /= len; }
 
     u[0] = s[1] * f[2] - s[2] * f[1];
@@ -291,7 +291,7 @@ void MatrixLookAt(float* m, float eyeX, float eyeY, float eyeZ,
 
 void MatrixPerspective(float* m, float fov, float aspect, float near, float far)
 {
-    float tanHalfFov = tan(fov / 360.0f * M_PI);
+    float tanHalfFov = (float)tan(fov / 360.0f * M_PI);
     int i;
     for (i = 0; i < 16; i++) m[i] = 0.0f;
     m[0] = 1.0f / (aspect * tanHalfFov);
